[Feature] Meta - Add the Meta functions (#22)
- Fix Authentication issues with DeviceFlow
  - Fixes #20 
  - Fixes #19
- The Auth module now supports Device Flow for both GitHubApp, OAuthApp,
PAT and System PAT.
    - If running Connect-GitHubAccount the flow will do the following:
      - If system tokens are available, use those to log in.
- Else trigger device flow. Check if access token is valid. if token is
valid for less than 4 hours, refresh the token using the device flow
refresh token process. Same if access token is expired.
- If refresh token is expired, trigger new device flow auth process.
- If type of auth changes due to switchs, trigger new device flow auth
process.
- If running Connect-GitHubAccount -AccessToken is run, an interactive
process is assumed. A prompt will ask the user to provide a PAT token,
which can be both Fine grained or classic.
- Changed so tokens are treated as SecureStrings from its stored until
its put in the function/command that uses it, and converts it there.

- Fixes #23 
- Configuration is now stored with a prefix that is configured in the
SecretVault.psd1 file in the data folder. Default this is "GHPS_". All
settings written to the secret store uses this.

- Adds an API call method `Invoke-GitHubAPI` to manage all the calls to
the GitHub API
- Includes default usage of settings, like ApiBaseUri, AccessToken (now
as a SecureString) and Version for the api version.
- Supports pagination, in line with [Using pagination in the REST
API](https://docs.github.com/en/rest/guides/using-pagination-in-the-rest-api?apiVersion=2022-11-28)
- Fixes #12 
  - Adds functions under the Meta namespace.