[Feature] Add Organization functions (#30)
- Fixes #29 
  - `private/Organization/Get-GitHubAllOrganization.ps1`
-
`public/Organization/Blocking/Assert-GitHubOrganizationBlockedUser.ps1`
  - `public/Organization/Blocking/Get-GitHubOrganizationBlockedUser.ps1`
  - `public/Organization/Get-GitHubOrganization.ps1`
  - `public/Organization/Get-GitHubOrganizationAppInstallation.ps1`
  - `public/Organization/Remove-GitHubOrganization.ps1`
  - `public/Organization/Set-GitHubOrganization.ps1`
  - `public/Organization/Set-GitHubOrganizationSecurityFeature.ps1`

Other fixes:
- `public/API/Invoke-GitHubAPI.ps1`
- Added support to convert body to query params when running a GET
function.
- Moved the `response` into a response property to allow for a more
dynamic interaction between the API function and the specific API
endpoint functions.
- Added `statuscode`, `responseheader` and `request` into the object
that is returned to the specific API endpoint functions. Required
specifically for function that use the `statuscode` as a way to answer
if users are blocked or not on Org or user level. i.e. [Check if a user
is blocked by an
organization](https://docs.github.com/en/rest/orgs/blocking?apiVersion=2022-11-28#check-if-a-user-is-blocked-by-an-organization)
- Added encoding/charset to `Accept` and `Content-Type` parameters, to
support use of Unicode icons. Reuired as I was playing around with
`Set-GitHubUser`, and setting bio to `I :heart: PowerShell`.
- `public/Auth/Connect-GitHubAccount.ps1`
- Fixed an issue where the remaining timer on access token validity
showed as non padded intigers. They now always show HH:mm:ss. 2+2+2 ints
as strings padded with 0's.
- Fixed an issue where the access token was not getting reset when
validity reached its half-life.
- `Utilities/`
- Added utility to manage hashtables and merging of objects + casing.
This has been used for the Config component, as well as the Auth
component. Now it has been revisited due to simplifying the means of
passing parameters to API calls with `$body = $PSBoundParameters |
ConvertFrom-HashTable | ConvertTo-Hashtable -NameCaseStyle snake_case`
as all the cases for API calls seem to be snake_cased.
- Utilities will be moved to the `Utilities` module in a later version.
It is kept here for now to not rely on the dependency.
- `Users/`
- Added other get user functions and wrapped them into `Get-GitHubUser`.
Different function are accessed using switches on `Get-GitHubUser`. For
more info check the `Get-GitHubUser` function.
- Simplify docs to not point to English docs by removing the `/en/`
component in the links.
- Remove unnecessary use of `Write-Output`.
- Replace `function` with `filter`, to streamline pipelining between
functions. There is very little other doc on this than 'Its like
`function` where the base script block is treated as `end` block by
default. Filters base script block is `process`.