🚀 [Feature]: Auto manage IAT when logged in as a GitHub App (#189)
## Description

- Fixes #170
- Added function to get named and authenticated app, `Get-GitHubApp`.
- Added a private function to get parameters used by parent function.
Used to get the parameters passed to a GitHub function to determine what
IAT context to use. i.e. if "Owner" is "PSModule" on a GitHub function,
the Get-FunctionParameter that is called in Resolve-GitHubContext get
the grandparent parameter, and looks for a context that can be used
towards "PSModule", with matching ClientIDs.
- GitHubContext class, added a 'Type' parameter to determine if the type
is a 'App' 'Installation' or 'User'.
- Added `Assert-GitHubContext` to start building a "rule engine" for
functions. In the future it will be added to all GitHub functions to
determine if the context that is used has the needed type, permission
etc. For this PR it is used in the `Connect-GitHubApp` function as it
should only be run if the context is an `App`.
- Updated Set-GitHubContext to name IAT based context by the following
pattern `hostname/login/targetname`.
- Added ouput/passthru functionality to `Update-GitHubUserAccessToken`
that refreshed UATs. This to avoid it having to store it in a context
and then directly after having to go get it again from the context. And
cleaned up the call in `Invoke-GitHubAPI`.
- Updated output type of `New-GitHubAppInstallationAccessToken`

 
## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [x] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas