🩹 [Patch]: Align function parameters and logging (#207)
## Description

- Fixes 🩹 [Patch]: Cleanup debug/verbose logging #187
- Fixes Align default parameter handling #78

This pull request introduces several new features and improvements,
including new PowerShell filters for GitHub workflow runs, updates to
the coverage report, and enhancements to existing scripts. The most
important changes are summarized below:

### Main change:

- Update all functions to use a function for showing the execution path
to the function in begin and end blocks of the functions. The exec path
starts at the first non '<ScriptBlock>' and non-powershell script so it
doesn't drown out the logs with useless info. Main goal of this change
is to clearly see where we are in the function execution. These only
show if `-Debug` is provided.
- Aligned all functions to use a $body hashtable for inputs to the API
call, instead of having some strange function with automated case style
management. Plain and simple, use a @body hashtable, an $inputObject
variable for the calls to the API and the API call itself. Easy as that.
Likely more alignment to come via classes to drive consistency at a
later point.
- All functions should now have a begin process and end block. Debug
functions are omitted from this requirement.

### Function changes:
- Splitting `Get-GitHubWorkflowRun` into two new private functions and
referencing them in the public function.
  - `Get-GitHubWorkflowRunByRepo`
  - `Get-GitHubWorkflowRunByWorkflow`

### Workflow Enhancements:
* Added `Initialize-PSModule` step to the `Set-CoverageReport.yml`
workflow, to install MarkdownPS automatically before the script is run.

### Codebase Enhancements:
* Updated `Get-GitHubAppByName`, `Get-GitHubAuthenticatedApp`,
`Assert-GitHubContext`, `Remove-GitHubContext`, `Resolve-GitHubContext`,
and `Set-GitHubContext` scripts to use `Write-Debug` instead of
`Write-Verbose` for improved debugging.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas

---------

Co-authored-by: github-actions <41898282+github-actions@users.noreply.github.com>