🪲 [Fix]: Load existing config when starting a new session (#211)
## Description

This pull request includes changes to the initialization process of the
GitHub configuration. The main goal is to improve the handling of the
GitHub configuration by ensuring it is properly initialized and loaded
from defaults or context when necessary.

* Fixes #210 

Improvements to GitHub configuration initialization:

*
[`src/functions/private/Config/Initialize-GitHubConfig.ps1`](diffhunk://#diff-0bd6f61981cdae153b550552b394da11c794e8ad8ae819fbb7bb702022a02e5dL32-L50):
Refactored the process block to better handle forced initialization and
loading of the GitHub configuration from context or defaults. Added
checks to avoid reinitializing if the configuration is already in
memory.

Changes to default configuration handling:

*
[`src/variables/private/Config.ps1`](diffhunk://#diff-24f682ffe9ba06c3a7c6620f5c550ddc9eac8bdfb1aefe07961c41cf1a5e9552L18-R18):
Set the initial value of `Config` to `$null` instead of creating a new
`GitHubConfig` instance. This ensures that the configuration is loaded
properly during initialization.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [x] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas