🩹 [Patch]: Let `Connect-GitHubApp` be able to pass the context to pipeline (#223)
## Description

This pull request includes enhancements to the `Connect-GitHubApp`
function, adding a `PassThru` parameter, enhancing output messages, and
updating tests to validate the new functionality.

- Fixes #222 

Enhancements to `Connect-GitHubApp` function:

*
[`src/functions/public/Auth/Connect-GitHubApp.ps1`](diffhunk://#diff-7d1951ca779d73ee11b11db4ade6f33f8ea5fcf052324a72d2c8e83304eaa045R64-R67):
Added a `PassThru` parameter to pass the context object to the pipeline.
*
[`src/functions/public/Auth/Connect-GitHubApp.ps1`](diffhunk://#diff-7d1951ca779d73ee11b11db4ade6f33f8ea5fcf052324a72d2c8e83304eaa045L136-R145):
Enhanced output messages to include a green checkmark.

Updates to tests:

*
[`tests/GitHub.Tests.ps1`](diffhunk://#diff-0b1d9ba345a583adce874126c13d6edd3f789416bb9c4db5df1e18af3608554cL98-R101):
Updated the test for connecting to a GitHub App Installation to validate
the `PassThru` functionality and ensure the context is not null or
empty.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas