🩹 [Patch]: Improve error handling in Git configuration scripts (#257)
## Description

This pull request includes changes to improve error handling in the
`Get-GitHubGitConfig.ps1` and `Set-GitHubGitConfig.ps1` scripts. The
most important changes include replacing exceptions with warnings and
adding checks to ensure the commands are run inside a Git repository.

Error handling improvements:

*
[`src/functions/public/Git/Get-GitHubGitConfig.ps1`](diffhunk://#diff-dfb306c31ba449aae53bfc9d39801cb64924641ebf9f953e6eed74a02877ee40L28-R34):
Replaced the exception thrown when Git is not installed with a warning
message and added a check to ensure the command is run inside a Git
repository.
*
[`src/functions/public/Git/Set-GitHubGitConfig.ps1`](diffhunk://#diff-9b7705b192d0886bb8d7c2ee853a341fee0be08850a23e15955110e7105c803bL38-R44):
Replaced the exception thrown when Git is not installed with a warning
message and added a check to ensure the command is run inside a Git
repository.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas