⚠️ [Security]: Mask sensitive data from `git config --list` (#259)
## Description

This pull request involves significant updates to the
`Get-GitHubGitConfig.ps1` script to enhance its functionality and
improve code readability. The most important changes include adding a
parameter for specifying the scope of the git configuration, improving
error messages, and updating the way git configuration data is processed
and returned.

Enhancements to functionality:

*
[`src/functions/public/Git/Get-GitHubGitConfig.ps1`](diffhunk://#diff-dfb306c31ba449aae53bfc9d39801cb64924641ebf9f953e6eed74a02877ee40L16-R20):
Added a new parameter `$Scope` with validation to allow specifying
'local', 'global', or 'system' scope for the git configuration.

Improvements to code readability:

*
[`src/functions/public/Git/Get-GitHubGitConfig.ps1`](diffhunk://#diff-dfb306c31ba449aae53bfc9d39801cb64924641ebf9f953e6eed74a02877ee40L29-R33):
Changed verbose message to use single quotes for consistency.

Updates to data processing:

*
[`src/functions/public/Git/Get-GitHubGitConfig.ps1`](diffhunk://#diff-dfb306c31ba449aae53bfc9d39801cb64924641ebf9f953e6eed74a02877ee40L43-R63):
Replaced the old method of processing git configuration data with a more
efficient approach using `ConvertFrom-StringData` and a hashtable to
store and return the results. This change also includes masking
sensitive information found in the configuration.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [x] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas