🩹 [Patch]: `git config` output as `hashtable[]` (#260)
## Description

This pull request includes changes to the Update-ConverageReport
workflow and updates to the `Get-GitHubGitConfig` function to return a
hashtable array instead of a `pscustomobject`.

Changes to GitHub Actions workflow:

* Renamed `.github/workflows/Set-CoverageReport.yml` to
`.github/workflows/Update-CoverageReport.yml` and updated the job names
and script references accordingly.
[[1]](diffhunk://#diff-e6405cf30762034bf8d27ee149a940b9d5136faad852c26538e8fcee8cb09cf6L1-R1)
[[2]](diffhunk://#diff-e6405cf30762034bf8d27ee149a940b9d5136faad852c26538e8fcee8cb09cf6L13-R13)
[[3]](diffhunk://#diff-e6405cf30762034bf8d27ee149a940b9d5136faad852c26538e8fcee8cb09cf6L22-R25)

Updates to `Get-GitHubGitConfig.ps1` script:

* Changed the `OutputType` attribute from `[pscustomobject]` to
`[object[]]`.
* Improved error handling and logging when retrieving Git configuration,
and refactored the script to use an array of hashtables (key value
pairs) for configuration storage instead of a `pscustomobject`. This is
to support multiple values with the same name, which `git config`
supports.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas