🩹 [Patch]: Update environment functions documentation and remove unused environment class (#323)
## Description

This pull request includes significant changes to the handling of GitHub
environments in the PowerShell module. The most important changes
involve the removal of the `Environment` class and its related classes,
updates to the example outputs in several functions, and some
adjustments to the function signatures.

### Removal of Classes:

*
[`src/classes/public/Environment/GitHubEnvironment.ps1`](diffhunk://#diff-c657e9b9cbcd747ecab0a2c2a667080554a022f31307be6ba5eb745cf70630bfL1-L68):
Removed the `Environment`, `RequiredReviewers`, `BranchPolicy`, and
`DeploymentBranchPolicy` classes, which were used to model GitHub
environment data.

### Updates to Example Outputs:

*
[`src/functions/private/Environments/Get-GitHubEnvironmentByName.ps1`](diffhunk://#diff-281ec1cdd72ba78fe2f547115ac26085efe035f4c30b748e579a7b71c3de9882L16-R34):
Updated the example output to reflect the new format of environment
details, including additional fields like `id`, `node_id`, and
`protection_rules`.
*
[`src/functions/private/Environments/Get-GitHubEnvironmentList.ps1`](diffhunk://#diff-b7e6a0698de85126a8c0bf19e78924ae225511ad39ea20b05b2fcf6d262e82efL13-R31):
Updated the example output to show the new format for listing
environments, including fields like `id`, `node_id`, `protection_rules`,
and `deployment_branch_policy`.
*
[`src/functions/public/Environments/Get-GitHubEnvironment.ps1`](diffhunk://#diff-4ce9e4defaf029427d3215c1a0688b6bf11a705532e17295851e80b2e8948b0fL15-R54):
Updated the example output to reflect the new format of environment
details, similar to the changes made in `Get-GitHubEnvironmentByName`.

### Function Signature Adjustments:

*
[`src/functions/public/Environments/Remove-GitHubEnvironment.ps1`](diffhunk://#diff-7318eee8b888e260d6170db2abcae8cba8146173a0a7036f2756777ebc9fd5e8L13-R21):
Changed the output type of the `Remove-GitHubEnvironment` function from
`PSCustomObject` to `void`, indicating that it no longer returns an
object.
*
[`src/functions/public/Environments/Set-GitHubEnvironment.ps1`](diffhunk://#diff-0d4fc874a278df0f48aba7cc73b05f002c068d4006e2acb0bdd24ab3eb6574e2L18-R40):
Modified the example to use a hashtable for parameters, demonstrating a
more complex usage scenario with additional fields like `Reviewers` and
`DeploymentBranchPolicy`.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas