🚀 [Feature]: Class overhaul - Contructors doing the attribute mapping (#344)
## Description

This pull request introduces several changes to improve the structure
and functionality of the PowerShell classes and configurations. The most
important changes include adding new classes, extending existing
classes, and updating configuration files.

### Class Enhancements

- All classes now handle attribute mapping from the raw object that was
received from the API call to the property of the class. Doing this to
"centralize" and deduplicate the data transformation.
- Standardize how to transform object `ID`s. The standard is to follow
GitHub's REST API by using `id` as the main ID (now stored in `ID`) and
the `node_id` as the `NodeID`.
- All classes that now have an `ID` and `NodeID` property extends the
`GitHubNode` class.
- Classes that have strong typed properties (like bool, int, datetime,
etc) are all nullable to support different coverage of the class. Some
APIs return an object that is partial, others full, so a class needs to
support both, else we need duplicate classes for partials and full
objects based on the API that we are using. NOT doing that at this
point. Trying instead to create a class structure that will be reusable.
- The `GitHubNode` class is also extended via Types to add an alias for
`ID` called `DatabaseID`.
- Some general class changes:
- `GitHubWebhook` and `GitHubWebhookRedelivery` is merged into
GitHubWebhookDelivery.
- Added classes for `GitHubLicense`, `GitHubPlan`, `GitHubOwner`,
`GitHubUser`, `GitHubOrganization`, `GitHubPermissions` and
`GitHubRepository`.

### Tests:

- Fix an issue where failed tests did not cleanup deployed repositories.
Also align repo cleanup after tests are complete.

### Configuration Updates

*
[`.github/linters/.markdown-lint.yml`](diffhunk://#diff-2fd3619d2b7372123257bc21645edb90c24293e66babbc6053033a89e21bab34R16):
Disabled rule `MD024` to allow duplicate headings for `INPUTS` and
`OUTPUTS`, allowing a filter function to both take a input type of
object and output the changed object or allow use of `PassThru`.

### Documentation Updates

* `CodingStandard.md`: Updated guidelines for class properties and added
new rules for using `GitHubNode` and property aliases.


## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [x] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas

---------

Co-authored-by: github-actions <41898282+github-actions@users.noreply.github.com>