🚀 [Feature]: Expand capability of `Repository` resources (#371)
## Description

This pull request expands the capability the module has on `Repository`
related resources by using a mix of GraphQL and REST APIs. This change
involves new concepts of mapping properties and fields that are
introduced by the new object schema of GraphQL object results. Other
than adding a new range of functionality for managing `Repository`
resources, we also add some QoL improvements like argument completion
for `Repository` and `Organization`.

### `Repository` changes

#### Classes

##### `GitHubLicenseRule`

* Introduced a new `GitHubLicenseRule` class to encapsulate license rule
details such as description, key, and label.

##### `GitHubRepositoryPermissions`

* Added a constructor to the `GitHubRepositoryPermissions` class to
initialize permissions based on a single string input, simplifying
permission handling logic.

##### `GitHubLicense`

* Added a new property for the customary shortname of a license type.
* Removed the `ApiUrl` property as we cover access to the license via
functions.
* Updated the properties `Permissions`, `Conditions`, and `Limitations`
to use the `GitHubLicenseRule` class.

##### `GitHubRepositoryLanguage`

- Introduced a new `GitHubRepositoryLanguage` to cover repo languages.
This stores color and name so that if we want to make some 'eye-candy'
later, we can.

##### `GitHubRepository`

* Removed property `IsDisabled` as there seem to be no use for this
property and no documentation what it does. There is also no place in
the GitHub UI where this is exposed or changable.
* Added `ArchivedAt` property to be able to track when a repo was
archived.
* Updated the `Language` property to use the new
`GitHubRepositoryLanguage` class.
* Added `OpenPullRequests` as a counter for how many PRs is open on the
repo.
* Renamed `AllowUpdateBranch` to `SuggestUpdateBranch` to reflect what
it actually does. It suggests to a user to update the branch in the PR
UI.
* Added a static class function to map class properties to GraphQL query
fields.
* Updated class constructor to have separate logic for GraphQL and REST
API result schemas, centralizing the mapping logic and keeping functions
"clean".

#### Formats

##### `GitHubRepository`

* Fix the `Size` format to use string formatter, round operator and
padding of 2 fractions.

#### Functions

- Reintroducing the `Owner` parameter, to correctly get a repo under
either a user or an organization without having to specify.
- 🌟 _Dev note: Mostly the reason of this update, but it just kept
growing!_
- Removing sorting options, assuming it is unnecessary for PowerShell
based GitHub integrations. This comes built in.
- Added `Property` and `AdditionalProperty` parameters to repository
query functions for flexible field selection in GraphQL queries.
- `Property` has a default for what it includes, if you want to override
these use this parameter.
- Use the `AdditionalProperty` parameter if you want more info than the
base properties.
- GraphQL-based functions also use the class based function to map the
requested class properties to GraphQL fields.
- Some parameters that was `switch` type, is now `bool` type.
- Removed reliance on some deprecated controls, like the `private`
property and RepoTeams.
- Removed functions that ends with `*TagProtection` as these are
deprecated from GitHub's side. Removing them without adding the
supported approach called `Rules` or `RuleSets`. This will come on a
later time.

##### Get-GitHubRepository

- Changed to use the new GraphQL-based private functions. New parameters
to play with.

##### New-GitHubRepository

- Overhauled to use private REST-based functions to create repo for
user, organization, from fork or from template with minimal settings in
that call, while offloading settings and config to
`Update-GitHubRepository`. This is due to the REST function having lack
of settings for some of these calls, and incomplete on others.

##### Update-GitHubRepository

- Overhauled to use REST and GraphQL APIs as these are partially
implemented in both places. See #354 for details.

#### Completers

- Added completers for `Get-GitHubRepository` for the parameters
`Property` and `AdditionalProperty`, so users can easily find available
options.
- Added completers globally for the `Repository` parameter. Will bind to
the `Owner` or `Organization` parameters to scope the returned
suggestions.
- Added completers for all `*GithubRepository` functions for the `Name`
parameter. This so that users can choose repos by name instead of copy
pasting.

### GraphQL capabilities

* Added a private function (`ConvertTo-GitHubGraphQLField`) that maps
class properties to GraphQL query fields.

### GitHub Status

- Update so that GitHub Status functions now support `Public`,
`Europe`(was `eu`) and `Australia`(new).
- Updated tests to test all of these.

### Disabled the 'Coverage Report' functionality

* The 'Coverage Report' now gives little value and overview of what the
module is capable of as we start using more a mix of GraphQL and REST
APIs. It would need to be object and action based, i.e. verb/noun list
per resource that is available in the platform. We will address this
issue separately, but for now disabling it.

### Refactor pagination parameter validation across multiple scripts

- Changed the type of `$PerPage` to `[System.Nullable[int]]` to allow
for optional usage.
- Removed unnecessary body creation for API calls, directly passing the
`PerPage` parameter to the input object.
- Ensured consistent handling of the `PerPage` parameter across user,
organization, webhook, and repository related scripts.

### Other

* Fixed some typos.
* Fix the `Size` format on `GitHubArtifact` to use string formatter,
round operator and padding of 2 fractions. Aligned with what is done
with `GitHubRepository`.
* All `Visibility` parameters are now capitalized, with a function that
takes them to lowercase before calling the API.
* Added test cleanup functionality to remove repositories they are using
before the tests are being run. It is common that previous tests failed
due to rate limits, causing them not to be able to clean up after the
tests. Applies to repositories created during the following tests:
Environments, Releases, Repositories, Secrets, Variables,
* Added argument completers for `Organization` functions.
* Functions called `*GitHubOrganization` now has completers for the
`Name` parameter.
  * Globally added a completer for the parameter `Owner`.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [x] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas