🩹 [Patch]: Improve `Secret` and `Variable` functions (#379)
## Description

This pull request introduces a new format for `GitHubTeam` objects and
replaces `.LINK` tags with `.NOTES` tags in multiple PowerShell script
files to enhance documentation consistency, adds many completers for
different resources and cleans up some functions and tests for
`Variables`, `Secrets` and `Repositories`.

### New Feature: Table View for `GitHubTeam`

* Added an XML configuration file, `GitHubTeam.Format.ps1xml`, to define
a table view for `GitHubTeam` objects. The table includes columns for
`Name`, `Owner`, and `Url`. This improves the display of `GitHubTeam`
data in PowerShell.

### Documentation Updates: `.LINK` to `.NOTES` Conversion

* Replaced `.LINK` tags with `.NOTES` tags across various PowerShell
functions to align with documentation standards. Each `.NOTES` tag now
contains a link to the relevant GitHub REST API documentation.

### Parameter completers

Added parameter completers for names for:
- Environments
- Secrets
- Teams
- Variables
- Workflows

### Pipeline support

Added pipeline support for:
- Secrets
- Teams
- Variables

### Test cleanup

Fix cleanup of tests for:
- Secrets
- Teams
- Variables

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas